3.2.1 refactoring naming modules - исправил имена файлов модулей, согласно замечанию наставника
