+++ src/view/film-board.js
